// src/shared/types/PermissionRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Permission } from "./Permission";
import type { PermissionStatus } from "./PermissionStatus";

export type PermissionRow = { permission: Permission, status: PermissionStatus, rationale: string, settings_url: string, };